requirements.txt
alembic==1.18.3
anyascii==0.3.3
anyio==4.12.1
astor==0.8.1
attrdict==2.0.1
attrs==25.4.0
babel==2.18.0
bce-python-sdk==0.9.60
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cssselect==1.4.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
et_xmlfile==2.0.0
exceptiongroup==1.3.1
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
future==1.0.0
greenlet==3.3.1
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langdetect==1.0.9
lazy_loader==0.4
llvmlite==0.46.0
lmdb==1.7.5
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.4.2
num2words==0.5.14
numba==0.64.0
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==26.0
paddleocr==2.7.0.3
paddlepaddle==2.6.2
pandas==2.3.3
pdf2docx==0.5.8
pillow==12.1.0
platformdirs==4.9.2
pooch==1.9.0
premailer==3.10.0
protobuf==3.20.2
psutil==7.2.2
pyclipper==1.4.0
pycryptodome==3.23.0
PyMatting==1.1.15
PyMuPDF==1.20.2
pyparsing==3.3.2
PyPDF2==3.0.1
pypdfium2==5.4.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-doctr==1.0.1
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
rarfile==4.2
referencing==0.37.0
rembg==2.0.69
reportlab==4.4.9
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
sympy==1.14.0
termcolor==3.3.0
tifffile==2025.5.10
tomli==2.4.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
validators==0.35.0
visualdl==2.5.3
Werkzeug==3.1.5
